Move session id as primary output of invoke agent operation
