+++ src/main/java/com/mulesoft/connector/agentforce/internal/botapi/group/BotAgentParameterGroup.java
@@ -1,6 +1,8 @@
 package com.mulesoft.connector.agentforce.internal.botapi.group;
 
 import com.mulesoft.connector.agentforce.internal.botapi.metadata.AgentListValueProvider;
+import org.mule.runtime.api.meta.ExpressionSupport;
+import org.mule.runtime.extension.api.annotation.Expression;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.param.display.DisplayName;
 import org.mule.runtime.extension.api.annotation.param.display.Placement;
@@ -11,6 +13,7 @@ public class BotAgentParameterGroup {
   @Parameter
   @Placement(order = 1)
   @OfValues(AgentListValueProvider.class)
+  @Expression(value = ExpressionSupport.NOT_SUPPORTED)
   @DisplayName("Agent List")
   private String agent;
